#56 Update uniform sampler implementation
The distinct SamplerBase and UniformSampler classes are no more. There
is now only the Sampler class, which is templated on sampled space and
the sampling distribution. The UniformSampler class is a type alias on
a template specialization.

Creating sampled vectors is now different too. It is assumed that there
is a function to convert an array to the desired vector type. The R2
and example custom spaces have been updated to reflect these changes.