#67 Introdue new RRT function
This new function will soon replace the existing one. The new one has
a slightly different implementation. Instead of checking if a path to
target could be made before adding the target to the tree, it simply
adds target to the tree then searches for a path. If no path is found,
the target vertex gets removed, and the search continues. This is more
in line with the RRT algorithm's original design.

The rrt.cpp example code has been updated accordingly. It also now uses
Eigen::Vector2f as the R2 vector instead of the original tuple-based
one. The tuple-based vector will eventually be removed from the library
since it is not directly related to motion planning.