#49 Add orthotope, type traits, dimension bounds
The orthotope is a necessary component of the future sampling
implementation. It will be used as the bounds for uniform random
samplers. The dimension bounds class is a supporting data structure.
The dimensionality type trait provides information how the mathematical
vector's size. It is assumed to be compile-time known even if the
vector's storage implementation is not.