#49 Add new sampling implementation
The new sampling implementation uses a specific distribution. The
rationale behind this design change is that there is currently no need
to generically support different distributions. This change is motivated
by the desire to keep things simple. We can always go more general if a
usage or implementation pattern emerges.