Add new sampler implementations (#50)
* #49 Add function to make mathematical vector

This function will be used in a future sampling class that will
replace the existing one. Having make_vector function is cleaner than
the existing resize and copy buffer strategies.

* #49 Add orthotope, type traits, dimension bounds

The orthotope is a necessary component of the future sampling
implementation. It will be used as the bounds for uniform random
samplers. The dimension bounds class is a supporting data structure.
The dimensionality type trait provides information how the mathematical
vector's size. It is assumed to be compile-time known even if the
vector's storage implementation is not.

* #49 Add new sampling implementation

The new sampling implementation uses a specific distribution. The
rationale behind this design change is that there is currently no need
to generically support different distributions. This change is motivated
by the desire to keep things simple. We can always go more general if a
usage or implementation pattern emerges.