Update the RRT implementation (#60)
* #53 Move attempt count class to separate file

Also remove the noexcept specifier for the sampler class.

* #53 Add Boost.Graph extensions

These are convenience type aliases, concepts, and exceptions that
supplement the vanilla Boost.Graph library for the
libwheel::motion_planning library.

* #53 Add new RRT implementation

This is still a work in progress because the nearest get_nearest_vertex
function still needs to be implemented.

* #53 Update RRT example

Since the RRT algorithm implementation is mostly done, we can go ahead
and update the example. The old one used the previous implementation
and needed to be gutted.