Generalize the RRT family of search algorithms (#69)
* #68 Add incremental sample and search algorithm

This implementation uses the same key points from the
RapidlyExploringRandomTrees class but with customization points. The
RapidlyExploringRandomTrees class will eventually be replaced.

* #68 Add simple RRT implementation

This implementation uses the template method design pattern and is
functionally equivalent to the RapidlyExploringRandomTrees class.

* #68 Remove old basic RRT implementation

The RapidlyExploringRandomTrees class was removed because it was fully
replaced by the SimpleRRT class. The example program has also been
updated.